+++ internal/recovery/kill_other.go
@@ -1,0 +1,7 @@
+//go:build !unix
+
+package recovery
+
+func killSelfAt(label CrashLabel) error {
+	return stateErr("KillAt unsupported on this OS")
+}